requirements.txt
fastapi
uvicorn
supabase
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
email-validator
